Fix ci test issue-94239.rs

+++ src/test/ui/issues/issue-94239.rs
@@ -1,6 +1,6 @@
-pub const fn test_match_range(len: usize) -> usize {
+pub const fn test_match_range(len: u64) -> u64 {
     match len {
-        10000000000000000000..=99999999999999999999 => 0, //~ ERROR literal out of range for `usize`
+        10000000000000000000..=99999999999999999999 => 0, //~ ERROR literal out of range for `u64`
         _ => unreachable!(),
     }
 }
